number of cluster should be an int ... you know...

+++ enjoycluster.py
@@ -14,7 +14,7 @@ def timeFeature(time):
 
 X_shifts = [[shift['a_lat'], shift['a_lon'], shift['b_lat'], shift['b_lon'], timeFeature(shift['a_time']) ] for shift in shifts]
 
-k_means = cluster.KMeans(n_clusters=math.sqrt(len(shifts)/2))
+k_means = cluster.KMeans(n_clusters=int(math.sqrt(len(shifts)/2)))
 k_means.fit(X_shifts)
